tests/submission/Microsoft/forms/inputwillvalidate.html

<!DOCTYPE HTML>
<html>
    <head>
        <title>willValidate property on the input element</title>
        <meta http-equiv="content-type" content="text/html; charset=UTF-8" />
        <meta name="description" content="willValidate property on the input element" />
    </head>
    <body>
        <script src="/resources/testharness.js" ></script>
        <script src="/resources/testharnessreport.js" ></script>

        <div id=log></div>
        
        <form style="display : none" action="http://www.example.com/">
          <input type="text" required></input>
          <button type="submit">Submit</button>
          <input type="text" disabled></input>
        </form>

        <script type="text/javascript">

        test(function() {assert_true(document.getElementsByTagName("input")[0].willValidate)}, "willValidate property returns true when required attribute exists");
        test(function() {assert_false(document.getElementsByTagName("input")[1].willValidate)}, "willValidate property returns false when disabled attribute exists");

        </script>
    </body>
</html>
